Enum vit_servicing_station_cli_f10::csv::loaders::CsvDataCmd

pub enum CsvDataCmd {
    Load {
        db_url: String,
        funds: PathBuf,
        voteplans: PathBuf,
        proposals: PathBuf,
        challenges: PathBuf,
        reviews: PathBuf,
        goals: PathBuf,
    },
}

Variants

Load

Fields

db_url: String

URL of the vit-servicing-station database to interact with

funds: PathBuf

Path to the csv containing funds information At the moment, it’s required these are ordered.

Also the first fund being the current one, which means previous funds should not be included. This restriction may be lifted in the future.

voteplans: PathBuf

Path to the csv containing voteplans information

proposals: PathBuf

Path to the csv containing proposals information

challenges: PathBuf

Path to the csv containing challenges information

reviews: PathBuf

Path to the csv containing advisor reviews information

goals: PathBuf

Path to the csv containing goals information

Load Funds, Voteplans and Proposals information into a SQLite3 ready file DB.

Implementations

impl CsvDataCmd

fn load_from_csv<T: DeserializeOwned>(csv_path: &Path) -> Result<Vec<T>>

fn handle_load( db_url: &str, funds_path: &Path, voteplans_path: &Path, proposals_path: &Path, challenges_path: &Path, reviews_path: &Path, goals_path: &Path ) -> Result<(), Error>

fn handle_load_with_db_backup( db_url: &str, funds_path: &Path, voteplans_path: &Path, proposals_path: &Path, challenges_path: &Path, reviews: &Path, goals: &Path ) -> Result<(), Error>